requirements.run.txt
Flask
loguru
pandas
scikit-learn
python-dotenv
sqlalchemy
psycopg2-binary